fix index.js 1.0
1. Перенес вынес весь код из index.js
2. Вынес селекторы в refs
