+++ src/js/login/login_close_window.js
