--- src/js/refsMarina.js
@@ -1,17 +1,0 @@
-
-
-
-export default {
-    // элементы хедера:
-    headerMain: document.getElementById('headerMain'),
-headerLib: document.getElementById('headerLib'),
-
-    btnHomeOpen: document.getElementById('btn-home-open'),
-    btnLibOpen: document.getElementById('btn-lib-open'),
-
-// кнопки в бібліотеці:
-    watchedBtn: document.getElementById('watched__btn'),
-    queueBtn: document.getElementById('queue__btn'),
-    //(хто буде підключати, то кнопки мають стилі на активний стан, властивість current)
-    
-}
